Move gremlin connector logic
