+++ packages/graph-explorer/src/connector/gremlin/__mock/edges-labels-and-counts.json
