+++ packages/graph-explorer/src/connector/gremlin/__mock/edges-schema.json
