+++ packages/graph-explorer/src/connector/gremlin/__mock/should-return-1-random-node.json
