+++ packages/graph-explorer/src/connector/gremlin/__mock/should-return-all-neighbors-from-node-2018.json
