+++ packages/graph-explorer/src/connector/gremlin/__mock/should-return-filtered-neighbors-from-node-2018.json
