+++ packages/graph-explorer/src/connector/gremlin/__mock/should-return-neighbors-counts-for-node-123.json
