+++ packages/graph-explorer/src/connector/gremlin/__mock/vertices-labels-and-counts.json
