+++ packages/graph-explorer/src/connector/gremlin/__mock/vertices-schema.json
