+++ packages/graph-explorer/src/connector/gremlin/fetchNeighbors/oneHopTemplate.ts
