+++ packages/graph-explorer/src/connector/gremlin/fetchNeighborsCount/index.ts
@@ -2,7 +2,7 @@ import type {
   NeighborsCountRequest,
   NeighborsCountResponse,
 } from "@/connector/useGEFetchTypes";
-import neighborsCountTemplate from "../templates/neighborsCountTemplate";
+import neighborsCountTemplate from "./neighborsCountTemplate";
 import type { GInt64 } from "../types";
 import { GremlinFetch } from "../types";
