+++ packages/graph-explorer/src/connector/gremlin/fetchNeighborsCount/neighborsCountTemplate.ts
