+++ packages/graph-explorer/src/connector/gremlin/fetchSchema/edgeLabelsTemplate.ts
