+++ packages/graph-explorer/src/connector/gremlin/fetchSchema/edgesSchemaTemplate.ts
