+++ packages/graph-explorer/src/connector/gremlin/fetchSchema/vertexLabelsTemplate.ts
