+++ packages/graph-explorer/src/connector/gremlin/fetchSchema/verticesSchemaTemplate.ts
