+++ packages/graph-explorer/src/connector/gremlin/fetchVertexCountsByType/vertexTypeCountTemplate.ts
