+++ packages/graph-explorer/src/connector/gremlin/keywordSearch/keywordSearchTemplate.ts
